+++ tests/test_F0_output.py
@@ -22,7 +22,17 @@ import numpy as np
 import src.B0_data_input_json as B0
 import src.F0_output as F0
 
-from .constants import TEST_REPO_PATH, DICT_PLOTS, PDF_REPORT
+from .constants import (
+    EXECUTE_TESTS_ON,
+    TESTS_ON_MASTER,
+    TEST_REPO_PATH,
+    DICT_PLOTS,
+    PDF_REPORT,
+    TYPE_DATETIMEINDEX,
+    TYPE_DATAFRAME,
+    TYPE_SERIES,
+    TYPE_TIMESTAMP,
+)
 from mvs_eland_tool.mvs_eland_tool import main
 import src.A0_initialization as initializing
 
@@ -36,6 +46,9 @@ VALUES = [0, 1, 2]
 
 pandas_DatetimeIndex = pd.date_range(start=START_TIME, periods=PERIODS, freq="60min")
 pandas_Series = pd.Series(VALUES, index=pandas_DatetimeIndex)
+pandas_Series_tuple_name = pd.Series(
+    VALUES, index=pandas_DatetimeIndex, name=(("A", "B"), "flow")
+)
 pandas_Dataframe = pd.DataFrame({"a": VALUES, "b": VALUES})
 SCALAR = 2
 
@@ -46,6 +59,7 @@ JSON_TEST_DICTIONARY = {
     "pandas_DatetimeIndex": pandas_DatetimeIndex,
     "pandas_Timestamp": pd.Timestamp(START_TIME),
     "pandas_series": pandas_Series,
+    "pandas_series_tuple_name": pandas_Series_tuple_name,
     "numpy_array": np.array(VALUES),
     "pandas_Dataframe": pandas_Dataframe,
 }
@@ -184,8 +198,9 @@ class TestFileCreation:
         assert os.path.exists(os.path.join(OUTPUT_PATH, file_name + ".json")) is True
 
     @pytest.mark.skipif(
-        "tests/{}".format(os.path.basename(__file__)) not in sys.argv,
-        reason="lengthy test",
+        EXECUTE_TESTS_ON not in (TESTS_ON_MASTER),
+        reason="Benchmark test deactivated, set env variable "
+        "EXECUTE_TESTS_ON to 'master' to run this test",
     )
     @mock.patch(
         "argparse.ArgumentParser.parse_args",
@@ -232,20 +247,22 @@ class TestDictionaryToJsonConversion:
         expr = F0.convert(JSON_TEST_DICTIONARY["pandas_DatetimeIndex"])
         assert (
             expr
-            == '{"columns":[0],"index":[1577836800000,1577840400000,1577844000000],"data":[[1577836800000],[1577840400000],[1577844000000]]}'
+            == TYPE_DATETIMEINDEX
+            + '{"columns":[0],"index":[1577836800000,1577840400000,1577844000000],"data":[[1577836800000],[1577840400000],[1577844000000]]}'
         )
 
     def test_processing_dict_for_json_export_parse_pandas_Timestamp(self):
         """ """
         expr = F0.convert(JSON_TEST_DICTIONARY["pandas_Timestamp"])
-        assert isinstance(expr, str)
+        assert expr == TYPE_TIMESTAMP + "2020-01-01 00:00:00"
 
     def test_processing_dict_for_json_export_parse_pandas_series(self):
         """ """
         expr = F0.convert(JSON_TEST_DICTIONARY["pandas_series"])
         assert (
             expr
-            == '{"name":null,"index":[1577836800000,1577840400000,1577844000000],"data":[0,1,2]}'
+            == TYPE_SERIES
+            + '{"name":null,"index":[1577836800000,1577840400000,1577844000000],"data":[0,1,2]}'
         )
 
     def test_processing_dict_for_json_export_parse_numpy_array(self):
@@ -257,15 +274,9 @@ class TestDictionaryToJsonConversion:
         """ """
         expr = F0.convert(JSON_TEST_DICTIONARY["pandas_Dataframe"])
         assert (
-            expr == '{"columns":["a","b"],"index":[0,1,2],"data":[[0,0],[1,1],[2,2]]}'
-        )
-
-    def test_processing_dict_for_json_export_parse_pandas_DatetimeIndex(self):
-        """ """
-        expr = F0.convert(JSON_TEST_DICTIONARY["pandas_DatetimeIndex"])
-        assert (
             expr
-            == '{"columns":[0],"index":[1577836800000,1577840400000,1577844000000],"data":[[1577836800000],[1577840400000],[1577844000000]]}'
+            == TYPE_DATAFRAME
+            + '{"columns":["a","b"],"index":[0,1,2],"data":[[0,0],[1,1],[2,2]]}'
         )
 
     def test_processing_dict_for_json_export_parse_unknown(self):
@@ -298,6 +309,11 @@ class TestLoadDictionaryFromJson:
         k = "pandas_series"
         assert self.value_dict[k].equals(JSON_TEST_DICTIONARY[k])
 
+    def test_load_json_parse_pandas_series_tuple_name(self):
+        """ """
+        k = "pandas_series_tuple_name"
+        assert self.value_dict[k].equals(JSON_TEST_DICTIONARY[k])
+
     def test_load_json_parse_numpy_array(self):
         """ """
         k = "numpy_array"
@@ -307,6 +323,16 @@ class TestLoadDictionaryFromJson:
         """ """
         k = "pandas_Dataframe"
         assert self.value_dict[k].equals(JSON_TEST_DICTIONARY[k])
+
+    def test_load_json_export_parse_pandas_DatatimeIndex(self):
+        """ """
+        k = "pandas_DatetimeIndex"
+        assert self.value_dict[k].equals(JSON_TEST_DICTIONARY[k])
+
+    def test_load_json_export_parse_pandas_Timestamp(self):
+        """ """
+        k = "pandas_Timestamp"
+        assert self.value_dict[k] == JSON_TEST_DICTIONARY[k]
 
     def teardown_class(self):
         """ """