+++ CHANGELOG.md
@@ -49,6 +49,12 @@ tipps for module building, and hint that units in the MVS are not checked (PR #2
 - Tests for storage for the module A1 (#299)
 - Benchmark test with only battery and grid (#302)
 - Possibility to save the report generated in F2 as a pdf (#284)
+- Test for module D2 (#147)
+- Possibility to run benchmark tests selectively and make sure they are all run on master branch
+ (#320)
+ - Test function names for E1 (#145)
+- Possibility to deploy the report of the results in a browser (#323)
+- Test for the module D0 (#150)
 
 ### Changed
 - Shore power randomization improved + amount of available docks can be chosen (#202)
@@ -65,10 +71,14 @@ tipps for module building, and hint that units in the MVS are not checked (PR #2
 - Call timeseries plot function for each bus (#278)
 - rename "storage" parameter in A1 and tests_A1 to "asset_is_a_storage"
 - Serialize the DataFrame and arrays into the json_with_results.json (#304)
-- Convert serialized DataFrame and arrays back into these types in the B0.load_json function (#304)
+- Convert serialized DataFrame and arrays back into these types in the B0.load_json function
+ (#304, #322, #326)
 - The input from the csv files produce the same json than the json file (#286)
+ - Move the CSS styling code to a style sheet (#317)
+ - Change the input data for creating the dataframes for generating the optimization and costs' tables from xlsx file to json (#317) 
  
 ### Removed
+- Removed parameter ´oemof_file_name´ from ´simulation_settings.csv´ (#150), as well as from all input files etc. The name is hardcoded now.
 
 ### Fixed
 - Rename "boolean" to "bool" in example json file (#214)
@@ -88,6 +98,7 @@ tipps for module building, and hint that units in the MVS are not checked (PR #2
  - Display SOC (#278)
  - Automatic update of the test coverage with coveralls.io (#307)
 - Logging message for maximumCap value (#310)
+- Create_app function in F0 for standalone execution (#317)
 
 ## [0.2.0] - 2020-03-13